Move a-b to PSR-4
