+++ tests/lib/Activity/ManagerTest.php
