+++ tests/lib/App/CodeChecker/CodeCheckerTest.php
