+++ tests/lib/App/CodeChecker/DeprecationCheckTest.php
