+++ tests/lib/App/CodeChecker/StrongComparisonCheckTest.php
