+++ tests/lib/App/DependencyAnalyzerTest.php
@@ -14,7 +14,7 @@ use OC\App\Platform;
 use OCP\IL10N;
 use Test\TestCase;
 
-class DependencyAnalyzer extends TestCase {
+class DependencyAnalyzerTest extends TestCase {
 
 	/** @var Platform */
 	private $platformMock;
