+++ tests/lib/AppFramework/Http/ContentSecurityPolicyTest.php
