+++ tests/lib/AppFramework/Http/EmptyContentSecurityPolicyTest.php
