+++ tests/lib/AppFramework/Http/RedirectResponseTest.php
