+++ tests/lib/AppFramework/Http/StreamResponseTest.php
