+++ tests/lib/AppFramework/Http/TemplateResponseTest.php
