+++ tests/lib/AppFramework/Middleware/MiddlewareDispatcherTest.php
