+++ tests/lib/AppFramework/Middleware/MiddlewareTest.php
