+++ tests/lib/AppFramework/Middleware/Security/CORSMiddlewareTest.php
