+++ tests/lib/AppFramework/Middleware/Security/SecurityMiddlewareTest.php
