+++ tests/lib/AppFramework/Middleware/SessionMiddlewareTest.php
