+++ tests/lib/AppFramework/Utility/SimpleContainerTest.php
