+++ tests/lib/Archive/TestBase.php
