+++ tests/lib/Archive/ZIPTest.php
