+++ tests/lib/Authentication/Token/DefaultTokenMapperTest.php
