+++ tests/lib/Authentication/Token/DefaultTokenProviderTest.php
